Issue 386 Fix High priority findbugs in MQ 5.1.1 - changes #4

+++ mq/main/bridge/bridge-jms/src/main/java/com/sun/messaging/bridge/service/jms/JMSBridge.java
@@ -1201,7 +1201,7 @@ public class JMSBridge {
         val = attrs.getProperty(JMSBridgeXMLConstant.CF.CONNECTATTEMPTINTERVAL,
                                 JMSBridgeXMLConstant.CF.CONNECTATTEMPTINTERVAL_DEFAULT);
         if (val != null) {
-            attemptInterval = Integer.parseInt(val);
+            attemptInterval = Long.parseLong(val);
         }
         if (attemptInterval < 0) attemptInterval = 0;
         attemptInterval = attemptInterval*1000;
